This pull request tries to optimize the store process by trimming down on the python access layer of SQLAlchemy.
Consists of two parts:
1. Only get the review status rules from the source code in the initial time of the report storage. After all the reports were added to the database, in a separate steps first get all applying review status rules from the server, then do a "bulk update" of all the newly stored reports, where a GUI based review status rule applies.
2. Cache the severity of the reports. Before this patch, a relatively expensive search of the checker labels were done, for every report being stored. Due to the nature of the analysis results, the number of reports greatly outweigh the number of type of reports (checkers), by caching this severity level can improve the store time.

Profiled the storage time of CodeChecker (master) before and after the modifications:
The reports folder was made with master clang + master cppcheck with master CodeChecker, with the --enable-all flag:
`Number of processed analyzer result files  | 1047 `
`Number of analyzer reports                        | 184471`



Reports were stored into empty databases.
On the labels, the colors are preserved for the sets. (in case the labels are hidden on the overview labels.)

Baseline measurements:
Mass storage time was 1823 secs.
![image](https://user-images.githubusercontent.com/23135800/203060146-81081ac7-893c-4dd2-989e-ec01a2890719.png)
The above label shows the full profile of the entire API call.
The below label shows the `__add_report` method. Please not the `checker_labels.py(severity)` call batch of 145 seconds.
![image](https://user-images.githubusercontent.com/23135800/203060506-5e4d4e99-1951-44a8-9fd8-3702cedc5306.png)
The following label shows the detailed view of the get_review_status call batch:
![image](https://user-images.githubusercontent.com/23135800/203062753-8f60c7e8-c852-4b24-8613-ec1823014f04.png)
Here the magenta bar represents the parsing of the source files (parse_codechecker_review_comment), the runtime of that was: 140 sec. It is faster than the previous release, due to fixes applied: #3777
The other call batches are python (and SQLA) overhead of the per report severity query.

Measurements after modification:
Mass storage time was 1313 secs
![image](https://user-images.githubusercontent.com/23135800/203060857-0ba765c1-bfcc-4d6f-be6a-3f1620cfed59.png)
Please note the absence of the get review status rules call batch on the above label.
The orange section after the report_file.py is the remnant of the slow review status retrieval logic. Parsing the file is still needed, (parse_codechecker_review_comment) 
![image](https://user-images.githubusercontent.com/23135800/203061600-81052ce8-9354-4ae7-ae6f-7bec4ec3a06f.png)

The last labels shows the impact of severity levels cache:
This is 1000x times faster (on large report counts)
![image](https://user-images.githubusercontent.com/23135800/203064379-0152fd00-5210-4c21-9127-0bb64e091c6a.png)
![image](https://user-images.githubusercontent.com/23135800/203064749-4c101dce-5c5a-4dd4-8a9b-67634a7ca5ee.png)

All in all I've managed to shave off ~28% percent of the run time. (1823s -> 1313s), Which can be pretty significant when the store times are scratching the one hour mark.

The above figures clearly show, that the same improvement could be done for the `__add_report` method. (Cache the new reports, and do a bulk update after the setup. With this approach the bulk update step of this patch could be incorporated into the bulk insert.)

An interesting (and needed in my opinion) measurement would be to do the same measurements with a database heavily populated with review status rules.