@vodorok 2023-03-20
This modification will affect the `idle in transaction` timeout settings of the PostgreSQL config. For larger report folders I would advise 15-30 min settings.